PYIC-1718: Rename the CriReturn lambda
