+++ lambdas/retrieve-cri-oauth-access-token/src/main/java/uk/gov/di/ipv/core/retrievecrioauthaccesstoken/RetrieveCriOauthAccessTokenHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.credentialissuerreturn;
+package uk.gov.di.ipv.core.retrievecrioauthaccesstoken;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
@@ -41,7 +41,7 @@ import java.text.ParseException;
 
 import static uk.gov.di.ipv.core.library.domain.VerifiableCredentialConstants.VC_CLAIM;
 
-public class CredentialIssuerReturnHandler
+public class RetrieveCriOauthAccessTokenHandler
         implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
     private static final Logger LOGGER = LogManager.getLogger();
     private static final String CRI_VALIDATE_ENDPOINT = "/journey/cri/validate/";
@@ -58,7 +58,7 @@ public class CredentialIssuerReturnHandler
 
     private String componentId;
 
-    public CredentialIssuerReturnHandler(
+    public RetrieveCriOauthAccessTokenHandler(
             CredentialIssuerService credentialIssuerService,
             ConfigurationService configurationService,
             IpvSessionService ipvSessionService,
@@ -72,7 +72,7 @@ public class CredentialIssuerReturnHandler
     }
 
     @ExcludeFromGeneratedCoverageReport
-    public CredentialIssuerReturnHandler() {
+    public RetrieveCriOauthAccessTokenHandler() {
         this.configurationService = new ConfigurationService();
         this.credentialIssuerService =
                 new CredentialIssuerService(
