+++ lambdas/retrieve-cri-oauth-access-token/src/main/resources/IpvLambdaJsonLayout.json
