+++ lambdas/retrieve-cri-oauth-access-token/src/main/resources/log4j2.yaml
