+++ settings.gradle
@@ -3,7 +3,7 @@ include "lib",
 		"lambdas",
 		"lambdas:issue-client-access-token",
 		"lambdas:sessionend",
-		"lambdas:credentialissuerreturn",
+		"lambdas:retrieve-cri-oauth-access-token",
 		"lambdas:get-credential-issuer-config",
 		"lambdas:evaluate-gpg45-scores",
 		"lambdas:issuedcredentials",
